Update README.md
In this ongoing project, team members are measured monthly on their productivity before appraisals. The purpose of this is to raise awareness of processors' position and spur competition to be the best. A major challenge will be to factor in the TAT as that is determined on an individual basis. The challenge is currently being looked at and will definitely be resolved.